requirements.txt
langchain
google-generativeai
pymysql
sqlalchemy
python-dotenv
streamlit
cryptography